add test for invalid
