--- src/tests/validate/practitioner_remote_server/practitioner_remote_server_profile_in_url_invalid.validate.test.js
@@ -1,107 +1,0 @@
-const invalidPractitionerResource = require('./fixtures/invalid_practitioner.json');
-
-const expectedInvalidPractitionerResponse = require('./expected/invalid_practitioner_response.json');
-
-const USCorePractitionerProfile = require('./fixtures/us_core_profile_practitioner.json');
-
-const {
-    commonBeforeEach,
-    commonAfterEach,
-    getHeaders,
-    createTestRequest,
-} = require('../../common');
-const {describe, beforeEach, afterEach, test} = require('@jest/globals');
-const {ConfigManager} = require('../../../utils/configManager');
-const nock = require('nock');
-
-const fhirValidationUrl = 'http://foo/fhir';
-
-class MockConfigManager extends ConfigManager {
-    get fhirValidationUrl() {
-        return fhirValidationUrl;
-    }
-}
-
-describe('Practitioner Update Tests', () => {
-    beforeEach(async () => {
-        await commonBeforeEach();
-    });
-
-    afterEach(async () => {
-        await commonAfterEach();
-    });
-    describe('Practitioner Validate', () => {
-        test('Valid resource profile in url not body', async () => {
-            const request = await createTestRequest((c) => {
-                c.register('configManager', () => new MockConfigManager());
-                return c;
-            });
-
-            const getProfileScope = nock('http://hl7.org')
-                .get('/fhir/us/core/StructureDefinition/us-core-practitioner')
-                .reply(200, USCorePractitionerProfile);
-
-            // http://foo/fhir/StructureDefinition
-            const uploadProfileScope = nock(`${fhirValidationUrl}`, {
-                reqheaders: {
-                    'accept-encoding': 'gzip, deflate',
-                    'accept': 'application/json',
-                    'content-type': 'application/fhir+json',
-                },
-            })
-                .post('/StructureDefinition', body => body.id === 'us-core-practitioner')
-                .reply(200, {});
-
-            const validationScope = nock(`${fhirValidationUrl}`, {
-                reqheaders: {
-                    'accept-encoding': 'gzip, deflate',
-                    'accept': 'application/json',
-                    'content-type': 'application/fhir+json'
-                },
-            })
-                .post(
-                    '/Practitioner/$validate?profile=http://hl7.org/fhir/us/core/StructureDefinition/us-core-practitioner',
-                    invalidPractitionerResource
-                )
-                .reply(200, {
-                        'resourceType': 'OperationOutcome',
-                        'issue': [
-                            {
-                                'severity': 'error',
-                                'code': 'processing',
-                                'details': {
-                                    'coding': [
-                                        {
-                                            'system': 'http://hl7.org/fhir/java-core-messageId',
-                                            'code': 'VALIDATION_VAL_PROFILE_UNKNOWN_NOT_POLICY'
-                                        }
-                                    ]
-                                },
-                                'diagnostics': "Profile reference 'http://hl7.org/fhir/us/core/StructureDefinition/us-core-patient' has not been checked because it is unknown, and the validator is set to not fetch unknown profiles",
-                                'location': [
-                                    'Patient.meta.profile[0]',
-                                    'Line 1, Col 2'
-                                ]
-                            }
-                        ]
-                    }
-                );
-
-            let resp = await request
-                .get('/4_0_0/Practitioner')
-                .set(getHeaders());
-            // noinspection JSUnresolvedFunction
-            expect(resp).toHaveResourceCount(0);
-
-            resp = await request
-                .post('/4_0_0/Practitioner/$validate?profile=http://hl7.org/fhir/us/core/StructureDefinition/us-core-practitioner')
-                .send(invalidPractitionerResource)
-                .set(getHeaders());
-            // noinspection JSUnresolvedFunction
-            expect(resp).toHaveResponse(expectedInvalidPractitionerResponse);
-            expect(getProfileScope.isDone()).toBeTruthy();
-            expect(uploadProfileScope.isDone()).toBeTruthy();
-            expect(validationScope.isDone()).toBeTruthy();
-        });
-    });
-});
